requirements.txt
Flask
nltk
sumy
python-docx